cmd/atlas/internal/sqlparse: move parsers out
